extract validation logic to model
